pyproject.toml
[project]
name = "stock_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.12.0",
    "finnhub-python>=2.4.24",
    "google-generativeai<=0.8.5",
    "langchain==0.3.13",
    "langchain-chroma==0.1.4",
    "langchain-community==0.3.13",
    "langchain-core>=0.3.56",
    "langchain-experimental==0.3.4",
    "langchain-google-community==2.0.3",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters==0.3.4",
    "langdetect==1.0.9",
    "langgraph>=0.5.4",
    "langgraph-checkpoint>=2.1.0",
    "langgraph-sdk==0.1.48",
    "langsmith==0.2.6",
    "pillow>=11.3.0",
    "matplotlib>=3.7.0",
    "pydantic>=2.10",
    "pydantic-core>=2.23.4",
    "pydantic-settings==2.7.0",
    "yfinance>=0.2.55",
    "pypdf>=5.4.0",
    "anthropic>=0.49.0",
    "gradio>=5.22.0",
    "httpx>=0.28.1",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ipython>=8.12.0,<9.0.0",
    "sendgrid>=6.12.5",
    "boto3>=1.40.55",
    "fredapi>=0.5.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
